pyproject.toml
# pyproject.toml
# Modern Python project configuration
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shopify-analytics"
version = "1.0.0"
description = "Shopify Product Data Analytics Dashboard"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "Flask==2.3.3",
    "Flask-SQLAlchemy==3.0.5",
    "psycopg2-binary==2.9.7",
    "python-dotenv==1.0.0",
    "pandas==2.1.1",
    "gunicorn==21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.2",
    "pytest-cov==4.1.0",
    "pytest-flask==1.2.0",
    "pytest-mock==3.11.1",
    "black==23.7.0",
    "flake8==6.0.0",
    "isort==5.12.0",
    "mypy==23.7.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_optional = true
